@@ -6,14 +6,14 @@
 
     <groupId>dev.revivalo</groupId>
     <artifactId>PlayerWarps</artifactId>
-    <version>3.3.2</version>
+    <version>4.0.0</version>
     <packaging>jar</packaging>
 
     <name>PlayerWarps</name>
 
     <description>Warps created by players for players</description>
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>21</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
@@ -102,6 +102,14 @@
                     </execution>
                 </executions>
             </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.4.1</version>
+                <configuration>
+                    <outputDirectory>C:/Users/42073/Desktop/mcserver/plugins</outputDirectory>
+                </configuration>
+            </plugin>
         </plugins>
         <resources>
             <resource>
@@ -154,10 +162,6 @@
             <url>https://repo.essentialsx.net/releases/</url>
         </repository>
         <repository>
-            <id>paper-repo</id>
-            <url>https://papermc.io/repo/repository/maven-public/</url>
-        </repository>
-        <repository>
             <id>bg-repo</id>
             <url>https://repo.bg-software.com/repository/api/</url>
         </repository>
@@ -175,8 +179,7 @@
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot-api</artifactId>
-            <version>1.16.5-R0.1-SNAPSHOT</version>
-<!--            <version>1.8.8-R0.1-SNAPSHOT</version>-->
+            <version>1.21-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
@@ -194,7 +197,7 @@
         <dependency>
             <groupId>com.tchristofferson</groupId>
             <artifactId>ConfigUpdater</artifactId>
-            <version>2.1-SNAPSHOT</version>
+            <version>2.2</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
@@ -218,7 +221,7 @@
         <dependency>
             <groupId>de.rapha149.signgui</groupId>
             <artifactId>signgui</artifactId>
-            <version>2.5.0</version>
+            <version>2.5.4</version>
             <scope>compile</scope>
         </dependency>
         <dependency>
@@ -303,14 +306,20 @@
         <dependency>
             <groupId>eu.athelion.territory</groupId>
             <artifactId>Territory</artifactId>
-            <version>1.0</version>
+            <version>1.2.1</version>
             <scope>system</scope>
-            <systemPath>${project.basedir}/libs/Territory-1.0.jar</systemPath>
+            <systemPath>${project.basedir}/libs/Territory-1.2.1.jar</systemPath>
         </dependency>
         <dependency>
             <groupId>com.github.GriefPrevention</groupId>
             <artifactId>GriefPrevention</artifactId>
             <version>16.18.2</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <groupId>com.arcaniax</groupId>
+            <artifactId>HeadDatabase-API</artifactId>
+            <version>1.3.2</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
