@@ -4,7 +4,8 @@ import dev.revivalo.playerwarps.category.CategoryManager;
 import dev.revivalo.playerwarps.commandmanager.command.PwarpMainCommand;
 import dev.revivalo.playerwarps.configuration.Data;
 import dev.revivalo.playerwarps.configuration.file.Config;
-import dev.revivalo.playerwarps.hook.HookManager;
+import dev.revivalo.playerwarps.hook.HookRegister;
+import dev.revivalo.playerwarps.menu.MenuRegister;
 import dev.revivalo.playerwarps.updatechecker.UpdateChecker;
 import dev.revivalo.playerwarps.updatechecker.UpdateNotificator;
 import dev.revivalo.playerwarps.user.UserHandler;
@@ -60,7 +61,7 @@ public final class PlayerWarpsPlugin extends JavaPlugin {
         new Metrics(this, 12061);
         new UserHandler(this);
 
-        HookManager.hook();
+        HookRegister.hook();
 
         if (Config.UPDATE_CHECKER.asBoolean()) {
             new UpdateChecker(this, 79089).getVersion(pluginVersion -> {
@@ -106,6 +107,8 @@ public final class PlayerWarpsPlugin extends JavaPlugin {
             }.runTaskTimerAsynchronously(this, intervalInTicks, intervalInTicks);
         }
 
+        MenuRegister.registerFiles();
+
         new UpdateNotificator();
     }
 
@@ -137,6 +140,10 @@ public final class PlayerWarpsPlugin extends JavaPlugin {
                 getLogger().log(Level.SEVERE, "Failed to copy resource " + resourcePath, e);
             }
         }
+    }
+
+    public void executeCommandAsConsole(String command) {
+        getServer().dispatchCommand(getServer().getConsoleSender(), command);
     }
 
     @Nullable
