@@ -25,11 +25,10 @@ public enum Config {
     WARP_VISIT_NOTIFICATION,
     TELEPORT_DELAY,
     ALLOW_COLORS_IN_WARP_DISPLAY_NAMES,
-    WARP_LISTING_MENU_SIZE,
-    CATEGORIES_MENU_SIZE,
-    WARP_MANAGE_MENU_SIZE,
-    CONFIRMATION_MENU_SIZE,
-    WARP_PRICE,
+    FREE_WARPS,
+    WARP_PRICING,
+    WARP_FIXED_PRICE,
+    WARP_CUSTOM_PRICE,
     DATE_FORMAT,
     SORT_BY,
     SELECTED_SORT,
@@ -51,6 +50,7 @@ public enum Config {
     SET_DESCRIPTION_CHARACTERS_LIMIT,
     SET_DISPLAY_NAME_FEE,
     SET_ADMISSION_FEE,
+    FEATURE_WARP_FEE,
     DEFAULT_WARP_ITEM,
     DEFAULT_LIMIT_SIZE,
     ALLOW_ACCEPT_TELEPORT_MENU,
@@ -59,8 +59,10 @@ public enum Config {
     DENY_ITEM,
     NO_WARP_FOUND_ITEM,
     INSUFFICIENT_PERMISSIONS_ITEM,
+    FEATURE_WARP_DURATION,
+    FEATURED_LIST_FULL,
     FAVORITE_WARPS_ITEM,
-    CATEGORIES_BACKGROUND_ITEM,
+    FREE_FEATURE_POSITION_ITEM,
     WARP_LIST_ITEM,
     SEARCH_WARP_ITEM,
     SORT_WARPS_ITEM,
