@@ -34,11 +34,16 @@ public enum Lang {
     WARP_SUBMITTED_FOR_APPROVAL,
     WARP_APPROVED,
     WARP_DENIED,
+    WARP_FEATURED,
+    WARP_ALREADY_FEATURED,
+    FEATURED_LIST_FULL,
 
     LIMIT_REACHED,
     LIMIT_REACHED_OTHER,
     NO_WARP_FOUND,
     NO_WARP_AT_POSITION,
+    FREE_FEATURE_POSITION,
+    FREE_FEATURE_POSITION_LORE,
     ALREADY_OWNING,
     ALPHABETICAL,
     LATEST,
@@ -55,6 +60,7 @@ public enum Lang {
     INSUFFICIENT_BALANCE_FOR_ACTION,
     INSUFFICIENT_BALANCE_TO_TELEPORT,
     WARP_ACCESS_BLOCKED,
+    TELEPORTATION_UNSAFE,
     WARP_REMOVED,
     WARP_REMOVED_WITH_REFUND,
     WARP_IN_DELETED_WORLD,
@@ -171,7 +177,8 @@ public enum Lang {
     PWARP_ACCESSIBILITY_LORE,
     SET_PRICE_LORE,
     CHANGE_OWNER_LORE,
-    RENAME_WARP_LORE;
+    RENAME_WARP_LORE,
+    TELEPORTATION_PROGRESS;
 
     private static final Map<String, String> messages = new HashMap<>();
     private static final Map<String, List<String>> lists = new HashMap<>();
