@@ -3,6 +3,7 @@ package dev.revivalo.playerwarps.hook.register;
 import com.earth2me.essentials.Essentials;
 import dev.revivalo.playerwarps.PlayerWarpsPlugin;
 import dev.revivalo.playerwarps.hook.Hook;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.logging.Level;
@@ -17,10 +18,15 @@ public class EssentialsHook extends Importable implements Hook<Essentials> {
     }
 
     @Override
+    public @NotNull String getName() {
+        return "Essentials";
+    }
+
+    @Override
     public void register() {
-        isHooked = isPluginEnabled("Essentials");
+        isHooked = isPluginEnabled();
         if (isHooked)
-            essentials = (Essentials) getPlugin("Essentials");
+            essentials = (Essentials) getPlugin();
     }
 
     @Override
@@ -37,7 +43,6 @@ public class EssentialsHook extends Importable implements Hook<Essentials> {
     public void importWarps() {
         for (String warpName : essentials.getWarps().getList()) {
             PlayerWarpsPlugin.get().getLogger().log(Level.INFO, warpName);
-            //PlayerWarpsPlugin.getWarpHandler().addWarp(new Warp(new HashMap<>()));
         }
     }
 }
