@@ -1,6 +1,7 @@
 package dev.revivalo.playerwarps.hook.register;
 
 import dev.revivalo.playerwarps.hook.Hook;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class OraxenHook implements Hook<Void> {
@@ -8,8 +9,13 @@ public class OraxenHook implements Hook<Void> {
     private boolean isHooked;
 
     @Override
+    public @NotNull String getName() {
+        return "Oraxen";
+    }
+
+    @Override
     public void register() {
-        isHooked = isPluginEnabled("Oraxen");
+        isHooked = isPluginEnabled();
     }
 
     @Override
