@@ -2,13 +2,20 @@ package dev.revivalo.playerwarps.hook.register;
 
 import com.sk89q.worldguard.WorldGuard;
 import dev.revivalo.playerwarps.hook.Hook;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class WorldGuardHook implements Hook<WorldGuard> {
     private boolean isHooked;
+
+    @Override
+    public @NotNull String getName() {
+        return "";
+    }
+
     @Override
     public void register() {
-        isHooked = isPluginEnabled("WorldGuard");
+        isHooked = isPluginEnabled();
     }
 
     @Override
