@@ -1,6 +1,8 @@
 package dev.revivalo.playerwarps.util;
 
 import java.text.DecimalFormat;
+import java.text.NumberFormat;
+import java.util.Locale;
 
 public final class NumberUtil {
     private static final DecimalFormat DECIMAL_FORMAT = new DecimalFormat("###,###,###,###");
@@ -14,7 +16,12 @@ public final class NumberUtil {
         return (double) tmp / factor;
     }
 
-    public static String formatNumber(int number) {
+    public static String formatNumber(long number) {
         return DECIMAL_FORMAT.format(number);
+    }
+
+    public static String format(long number) {
+        NumberFormat numberFormat = NumberFormat.getInstance(Locale.getDefault());
+        return numberFormat.format(number);
     }
 }
