@@ -1,7 +1,7 @@
 package dev.revivalo.playerwarps.warp.checker;
 
 import dev.revivalo.playerwarps.configuration.file.Lang;
-import dev.revivalo.playerwarps.hook.HookManager;
+import dev.revivalo.playerwarps.hook.HookRegister;
 import me.ryanhamshire.GriefPrevention.Claim;
 import me.ryanhamshire.GriefPrevention.GriefPrevention;
 import org.bukkit.Location;
@@ -11,7 +11,7 @@ public class GriefPreventationChecker implements Checker {
     @Override
     public boolean validate(Player player) {
         Location loc = player.getLocation();
-        GriefPrevention griefPrevention = HookManager.getGriefPreventionHook().getApi();
+        GriefPrevention griefPrevention = HookRegister.getGriefPreventionHook().getApi();
         Claim claim = griefPrevention.dataStore.getClaimAt(loc, true, null);
         if (claim == null) {
             return true;
